feat: migrate react-router to TanStack Router, decompose App.tsx, remove bun.zip
- Replaced react-router-dom with @tanstack/react-router (file-based routing)
- Added @tanstack/router-plugin for Vite auto-generation
- Decomposed monolithic App.tsx (1174 lines) into 14 focused modules
- Removed bun.zip (38MB) from git tracking
- Added SEO meta tags to index.html
- Added 404 catch-all route with styled UI
- Enabled auto code-splitting via TanStack Router plugin
- Updated biome.json to ignore generated routeTree.gen.ts
- Removed unused App.css template boilerplate
