+++ index.html
@@ -4,7 +4,12 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>idea.web.front</title>
+    <meta name="description" content="Ideathon 2026 — An innovation-focused event fostering critical thinking and problem-solving through AI in Academia at Amrita Vishwa Vidyapeetham." />
+    <meta name="theme-color" content="#000000" />
+    <meta property="og:title" content="Ideathon 2026 — AI in Academia | Amrita" />
+    <meta property="og:description" content="An innovation-focused event fostering critical thinking and problem-solving through structured stakeholder requirement analysis." />
+    <meta property="og:type" content="website" />
+    <title>Ideathon 2026 — AI in Academia | Amrita</title>
   </head>
   <body>
     <div id="root"></div>
