+++ package.json
@@ -15,14 +15,15 @@
 	},
 	"dependencies": {
 		"@tailwindcss/vite": "^4.2.1",
+		"@tanstack/react-router": "^1.163.3",
 		"react": "^19.2.0",
 		"react-dom": "^19.2.0",
-		"react-router-dom": "^7.13.1",
 		"tailwindcss": "^4.2.1",
 		"tailwindcss-animate": "^1.0.7"
 	},
 	"devDependencies": {
 		"@biomejs/biome": "2.4.4",
+		"@tanstack/router-plugin": "^1.164.0",
 		"@types/node": "^25.3.0",
 		"@types/react": "^19.2.7",
 		"@types/react-dom": "^19.2.3",
