+++ src/hooks/use-intersection-observer.ts
@@ -1,0 +1,20 @@
+import { useEffect, useRef, useState } from "react";
+
+export function useIntersectionObserver(options: IntersectionObserverInit = {}) {
+    const [isIntersecting, setIsIntersecting] = useState(false);
+    const ref = useRef<HTMLDivElement>(null);
+
+    useEffect(() => {
+        const observer = new IntersectionObserver(
+            ([entry]) => {
+                if (entry.isIntersecting) setIsIntersecting(true);
+            },
+            { threshold: 0.1, ...options },
+        );
+
+        if (ref.current) observer.observe(ref.current);
+        return () => observer.disconnect();
+    }, [options]);
+
+    return { ref, isIntersecting };
+}
